+++ packages/netlify-cms-backend-github/src/AuthenticationPage.js
@@ -16,8 +16,9 @@ const ForkApprovalContainer = styled.div`
 `;
 const ForkButtonsContainer = styled.div`
   display: flex;
-  flex-flow: row nowrap;
+  flex-flow: column nowrap;
   justify-content: space-around;
+  align-items: center;
 `;
 
 export default class GitHubAuthenticationPage extends React.Component {
@@ -109,7 +110,7 @@ export default class GitHubAuthenticationPage extends React.Component {
     if (requestingFork) {
       const { approveFork, refuseFork } = this.state;
       return {
-        renderPageContent: ({ LoginButton }) => (
+        renderPageContent: ({ LoginButton, TextButton, showAbortButton }) => (
           <ForkApprovalContainer>
             <p>
               Open Authoring is enabled: we need to use a fork on your github account. (If a fork
@@ -117,7 +118,9 @@ export default class GitHubAuthenticationPage extends React.Component {
             </p>
             <ForkButtonsContainer>
               <LoginButton onClick={approveFork}>Fork the repo</LoginButton>
-              <LoginButton onClick={refuseFork}>Don&#39;t fork the repo</LoginButton>
+              {showAbortButton && (
+                <TextButton onClick={refuseFork}>Don&#39;t fork the repo</TextButton>
+              )}
             </ForkButtonsContainer>
           </ForkApprovalContainer>
         ),